% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{gsea_olink_run}
\alias{gsea_olink_run}
\title{gsea_olink_run}
\usage{
gsea_olink_run(
  x,
  gs = gsea_sets,
  minGSSize = 2,
  contrastName = compname,
  save = TRUE,
  saveFolder
)
}
\arguments{
\item{x}{output from an Olink statistical test}

\item{gs}{Gene set collection, Default: gsea_sets}

\item{minGSSize}{minimum gene set size to be considered, Default: 2}

\item{contrastName}{Name of contrast for filename purposes}

\item{save}{save results, logical, Default: TRUE}

\item{saveFolder}{path to folder where results are to be save if save=TRUE}
}
\value{
returns a cluterProfiler result object, suitable for further visualization
}
\description{
Run GSEA analysis from the output of an Olink statistical test using the estimate as the ranking metric
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[tibble]{enframe}}
\code{\link[openxlsx]{write.xlsx}}
}
