% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{gsea_olink_viz}
\alias{gsea_olink_viz}
\title{gsea_olink_viz}
\usage{
gsea_olink_viz(gsea_res, saveFolder, nterms = 10, contrastName = compname)
}
\arguments{
\item{gsea_res}{clusterProfiler result object}

\item{saveFolder}{path to folder where results are to be saved}

\item{nterms}{Number of terms to be visuzlized for several visualizations, Default: 10}

\item{contrastName}{Name of contrast for filename purposes , Default: compname}
}
\value{
Several visualizations are saved to disk in the folder provided
}
\description{
Create several visualizations starting from a cluterProfiler result object
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[enrichplot]{c("dotplot,compareClusterResult-method", "dotplot", "dotplot")}},\code{\link[enrichplot]{c("pairwise_termsim", "pairwise_termsim")}},\code{\link[enrichplot]{c("emapplot", "emapplot")}},\code{\link[enrichplot]{c("gseaplot2", "gseaplot2")}}
\code{\link[grid]{arrow}}
}
