% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{gsea_sets}
\alias{gsea_sets}
\title{GSEA Gene Set Collection}
\format{
A data frame with 1214439 rows and 2 variables:
\describe{
\item{\code{term}}{character Gene Set Name}
\item{\code{gene}}{character Gene}
}
}
\usage{
gsea_sets
}
\description{
Gene set collection from the Bader Lab, minimal clean up to provide better term names
}
\details{
DETAILS
}
\keyword{datasets}
