% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{olink_dist}
\alias{olink_dist}
\title{Olink Dist Plot}
\usage{
olink_dist(npx, panel, filename, width = 7, height = 7)
}
\arguments{
\item{npx}{NPX dataset (output from OlinkAnalyze::read_NPX)}

\item{panel}{Panel Name}

\item{filename}{filename, the suffix will determine the filetype}

\item{width}{Figure width,  Default: 7}

\item{height}{Figure height,  Default: 7}
}
\value{
distance plot
}
\description{
Create a Distance plot
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 olink_dist(plasma_npx, "Olink CARDIOVASCULAR III", filename="test.png")
 }
}
}
\seealso{
\code{\link[OlinkAnalyze]{olink_dist_plot}}
}
