% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{plasma_npx}
\alias{plasma_npx}
\title{plasma_npx}
\format{
A data frame with 54832 rows and 13 variables:
\describe{
\item{\code{SampleID}}{character Sample ID; matches metadata of plasma}
\item{\code{Index}}{integer Sample numbering}
\item{\code{OlinkID}}{character OlinkID}
\item{\code{UniProt}}{character UniProc code for the protein}
\item{\code{Assay}}{character Protein Symbol}
\item{\code{MissingFreq}}{character Frequency of missingness}
\item{\code{Panel}}{character Panel Name}
\item{\code{Panel_Version}}{character Paner Version}
\item{\code{PlateID}}{character Plate ID}
\item{\code{QC_Warning}}{character Is there a QC waringn?}
\item{\code{LOD}}{double Limit of Detection}
\item{\code{NPX}}{double NPX value; abundance measurement}
\item{\code{Normalization}}{character Normalization method}
}
}
\usage{
plasma_npx
}
\description{
NPX data of the plasma samples
}
\details{
DETAILS
}
\keyword{datasets}
